Move block related marshalling functions into their own source file
