+++ pkg/core/block/header.go
@@ -2,10 +2,7 @@ package block
 
 import (
 	"bytes"
-	"encoding/binary"
-	"io"
 
-	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire/encoding"
 	"github.com/dusk-network/dusk-crypto/hash"
 )
 
@@ -15,88 +12,6 @@ const (
 	// HeightSize size of a block height field in bytes
 	HeightSize = 8
 )
-
-func MarshalHashable(r io.Writer, h *Header) error {
-	if err := encoding.WriteUint8(r, h.Version); err != nil {
-		return err
-	}
-
-	if err := encoding.WriteUint64(r, binary.LittleEndian, h.Height); err != nil {
-		return err
-	}
-
-	if err := encoding.WriteUint64(r, binary.LittleEndian, uint64(h.Timestamp)); err != nil {
-		return err
-	}
-
-	if err := encoding.Write256(r, h.PrevBlockHash); err != nil {
-		return err
-	}
-
-	if err := encoding.WriteBLS(r, h.Seed); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func MarshalHeader(r io.Writer, h *Header) error {
-	if err := MarshalHashable(r, h); err != nil {
-		return err
-	}
-
-	if err := encoding.Write256(r, h.TxRoot); err != nil {
-		return err
-	}
-
-	if err := MarshalCertificate(r, h.Certificate); err != nil {
-		return err
-	}
-
-	if err := encoding.Write256(r, h.Hash); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func UnmarshalHeader(r io.Reader, h *Header) error {
-	if err := encoding.ReadUint8(r, &h.Version); err != nil {
-		return err
-	}
-
-	if err := encoding.ReadUint64(r, binary.LittleEndian, &h.Height); err != nil {
-		return err
-	}
-
-	var timestamp uint64
-	if err := encoding.ReadUint64(r, binary.LittleEndian, &timestamp); err != nil {
-		return err
-	}
-	h.Timestamp = int64(timestamp)
-
-	if err := encoding.Read256(r, &h.PrevBlockHash); err != nil {
-		return err
-	}
-
-	if err := encoding.ReadBLS(r, &h.Seed); err != nil {
-		return err
-	}
-
-	if err := encoding.Read256(r, &h.TxRoot); err != nil {
-		return err
-	}
-
-	if err := UnmarshalCertificate(r, h.Certificate); err != nil {
-		return err
-	}
-
-	if err := encoding.Read256(r, &h.Hash); err != nil {
-		return err
-	}
-
-	return nil
-}
 
 // Header defines a block header on a Dusk block.
 type Header struct {
